chore: when execute laravel init, stop clone project or install package if it's exist

+++ src/Commands/Laravel/Init.js
@@ -16,17 +16,55 @@ class LaravelInit extends BaseCommand{
 
     mkdir(deployConfig.path)
     try {
-      await execAsync(`git clone ${deployConfig.repo} ${deployConfig.name}`, { cwd: deployConfig.path })
-      await execAsync(`cp .env.example .env`, { cwd: laravelPath })
+      await this._cloneIfNotExist(deployConfig)
+      await this._createEnvFileIfNotExist(laravelPath)
       await execAsync(`composer install`, { cwd: laravelPath })
+      await this._installEnvSetCommandPackageIfNotExist(laravelPath)
       await execAsync(`php artisan key:generate`, { cwd: laravelPath })
       await execAsync(`php artisan storage:link`, { cwd: laravelPath })
       await execAsync(`chmod 755 -R ./`, { cwd: laravelPath })
-      await execAsync(`chmod -R o+w ./storage`, { cwd: laravelPath })
+      await this._createSymbolicLinkIfNotExist(laravelPath)
       await execAsync(`git config core.filemode false`, { cwd: laravelPath })
     } catch(error) {
       log(`Laravel init ${deployConfig.name} fail: ${error}`)
     }
+  }
+
+  async _cloneIfNotExist(deployConfig) {
+    const projectExist = this.existsSync(this.resolve(deployConfig.path, deployConfig.name))
+    if(projectExist) {
+      return
+    }
+
+    await execAsync(`git clone ${deployConfig.repo} ${deployConfig.name}`, { cwd: deployConfig.path })
+  }
+
+  async _createEnvFileIfNotExist(laravelPath) {
+    const envExist = this.existsSync(this.resolve(laravelPath, '.env'))
+    if(envExist) {
+      return
+    }
+
+    await execAsync(`cp .env.example .env`, { cwd: laravelPath })
+  }
+
+  async _installEnvSetCommandPackageIfNotExist(laravelPath) {
+    const composerJson = this.readFileSync(this.resolve(laravelPath, 'composer.json'), 'utf8')
+    const hasInstall = new RegExp(/imliam\/laravel-env-set-command/g).test(composerJson)
+    log(`hasInstall: ${hasInstall}`)
+    if(hasInstall) {
+      return
+    }
+
+    await execAsync(`composer require imliam/laravel-env-set-command:^1.0.0`, { cwd: laravelPath })
+  }
+
+  async _createSymbolicLinkIfNotExist(laravelPath) {
+    if(this.resolve(laravelPath, 'public/storage')) {
+      return
+    }
+
+    await execAsync(`chmod -R o+w ./storage`, { cwd: laravelPath })
   }
 }
